refactor(cache): split `memory_cache.rs` to several files
