+++ src/async_fuse/memfs/cache/memory_cache/mod.rs
@@ -1,0 +1,12 @@
+//! The in-memory cache.
+
+pub mod builder;
+mod cache;
+mod write_back_task;
+
+pub use builder::MemoryCacheBuilder;
+pub use cache::MemoryCache;
+pub use write_back_task::SoftLimit;
+
+#[cfg(test)]
+mod tests;
