+++ src/async_fuse/memfs/cache/mod.rs
@@ -8,16 +8,14 @@ mod global_cache;
 mod memory_cache;
 mod storage;
 mod storage_manager;
-mod write_back_task;
 
 pub mod policy;
 
 pub use block::{Block, BlockCoordinate};
 pub use global_cache::*;
-pub use memory_cache::{MemoryCache, MemoryCacheBuilder};
+pub use memory_cache::{MemoryCache, MemoryCacheBuilder, SoftLimit};
 pub use storage::Storage;
 pub use storage_manager::StorageManager;
-pub use write_back_task::{Command, SoftLimit};
 
 /// The number of bytes in one KiB.
 pub const KB_SIZE: usize = 1024;
